--- src/app/interfaces/acl-flags.ts
@@ -1,10 +1,0 @@
-export enum AclFlag {
-    FileInherit = 'FILE_INHERIT',
-    DirectoryInherit = 'DIRECTORY_INHERIT',
-    NoPropagateInherit = 'NO_PROPAGATE_INHERIT',
-    InheritOnly = 'INHERIT_ONLY',
-    Inherited = 'INHERITED',
-    Basic = 'BASIC'
-}
-
-export type AclFlags = { [K in AclFlag]: boolean };
